fix: フェーズを分離 - createStatementData をファイルに分割
